Parse forms in a page.

Parse forms in a page.

html_form(x)

Arguments

x
A node, node set or document.

See also

Examples

html_form(read_html("https://hadley.wufoo.com/forms/libraryrequire-quiz/"))

#> [[1]]
#> <form> 'form27' (POST https://hadley.wufoo.com/forms/libraryrequire-quiz/)
#>   <input hidden> 'Field1': 
#>   <input radio> 'Field1': Loads the class package
#>   <input radio> 'Field1': Loads the cluster package
#>   <input radio> 'Field1': I have no idea!
#>   <input radio> 'Field1': Something else
#>   <input text> 'Field1_other_Something else': 
#>   <input hidden> 'currentPage': GfqNaMDZduwwhzuA9YxErPxIPfa0GJOeczQWDlLzZzI=
#>   <input submit> 'saveForm': Submit
#>   <textarea> 'comment' [0 char]
#>   <input hidden> 'idstamp': K2usAn/awmxbltuyp6a77nxaxkLBqab+50PUZHtuuMI=
#>   <input hidden> 'stats': {"errors":0,"startTime":0,"endTime":0,"referer":null}
#>   <input hidden> 'clickOrEnter': 
#> 

html_form(read_html("https://hadley.wufoo.com/forms/r-journal-submission/"))

#> [[1]]
#> <form> 'form26' (POST https://hadley.wufoo.com/forms/r-journal-submission/)
#>   <input text> 'Field9': 
#>   <input email> 'Field11': 
#>   <textarea> 'Field13' [0 char]
#>   <input text> 'Field15': 
#>   <input file> 'Field17': 
#>   <input checkbox> 'Field19': I have read the <a href="https://github.com/rjournal/share/raw/master/author-guide.pdf">instructions for authors</a>
#>   <input checkbox> 'Field25': I am using version 0.12 of the style file
#>   <input checkbox> 'Field20': My bibliography is stored in a separate bibtex (.bib) file
#>   <input checkbox> 'Field24': My article is called lastname.tex and I have updated RJwrapper.tex
#>   <input checkbox> 'Field21': I acknowledge that if accepted my article will be published with CC-BY license
#>   <input checkbox> 'Field22': The manuscript has not been published in a peer reviewed journal, and is not currently under review for another journal
#>   <input checkbox> 'Field23': All authors agree to submission of the article to the R Journal
#>   <input text> 'Field119': 
#>   <input hidden> 'currentPage': GfqNaMDZduwwhzuA9YxErPxIPfa0GJOeczQWDlLzZzI=
#>   <input submit> 'saveForm': Submit
#>   <textarea> 'comment' [0 char]
#>   <input hidden> 'idstamp': FiPMOH2nE7N/7GidcHGwqzlYOaJ5HyGp2Mo/452uiEc=
#>   <input hidden> 'stats': {"errors":0,"startTime":0,"endTime":0,"referer":null}
#>   <input hidden> 'clickOrEnter': 
#> 

box_office <- read_html("http://www.boxofficemojo.com/movies/?id=ateam.htm")
box_office %>% html_node("form") %>% html_form()

#> <form> 'searchbox' (POST /search/q.php)
#>   <input text> 'q': 
#>   <input submit> '': Search...