jump_to {rvest}	R Documentation

Navigate to a new url.

Description

jump_to() takes a url (either relative or absolute); follow_link takes an expression that refers to a link (an <a> tag) on the current page.

Usage

jump_to(x, url, ...)

follow_link(x, i, css, xpath, ...)

Arguments

x
A session.

url
A URL, either relative or absolute, to navigate to.

...
Any additional httr configs to apply to this request.

i
You can select with:

an integer

selects the ith link

a string

first link containing that text (case sensitive)

css
Nodes to select. Supply one of css or xpath depending on whether you want to use a css or xpath 1.0 selector.

xpath
Nodes to select. Supply one of css or xpath depending on whether you want to use a css or xpath 1.0 selector.

Examples

s <- html_session("http://hadley.nz")
s <- s %>% follow_link("github")
s <- s %>% back()
s %>% follow_link("readr")

[Package rvest version 0.3.2 Index]