submit_form {rvest}	R Documentation

Submit a form back to the server.

Description

Submit a form back to the server.

Usage

submit_form(session, form, submit = NULL, ...)

Arguments

session
Session to submit form to.

form
Form to submit

submit
Name of submit button to use. If not supplied, defaults to first submission button on the form (with a message).

...
Additional arguments passed on to GET() or POST()

Value

If successful, the parsed html response. Throws an error if http request fails. To access other elements of response, construct it yourself using the elements returned by submit_request.

Examples

test <- google_form("1M9B8DsYNFyDjpwSK6ur_bZf8Rv_04ma3rmaaBiveoUI")
f0 <- html_form(test)[[1]]
f1 <- set_values(f0, entry.564397473 = "abc")

[Package rvest version 0.3.2 Index]