xml {rvest}	R Documentation

Work with xml.

Description

Deprecated. Please use just xml2 directly

Usage

xml(x, ..., encoding = "")

xml_tag(x)

xml_node(x, css, xpath)

xml_nodes(x, css, xpath)

Arguments

x
A url, a local path, a string containing html, or a response from an httr request.

...
If x is a URL, additional arguments are passed on to GET().

encoding
Specify encoding of document. See iconvlist() for complete list. If you have problems determining the correct encoding, try stri_enc_detect

css
Nodes to select. Supply one of css or xpath depending on whether you want to use a css or xpath 1.0 selector.

xpath
Nodes to select. Supply one of css or xpath depending on whether you want to use a css or xpath 1.0 selector.

[Package rvest version 0.3.2 Index]