Parse forms and set values

Use html_form() to extract a form, set values with html_form_set(), then submit it with session_submit()

html_form(x)

html_form_set(form, ...)

Arguments

x
A node, node set or document.

form
A form

...
<dynamic-dots> Name-value pairs giving fields to modify.

Provide a character vector to set multiple checkboxes in a set or select multiple values from a multi-select.

Value

An an S3 object with class rvest_form.

See also

Examples

session <- html_session("http://www.google.com")
search <- html_form(session)[[1]]

search <- search %>% html_form_set(q = "My little pony", hl = "fr")

#> Warning: Setting value of hidden field 'hl'.

# Or if you have a list of values, use !!!
vals <- list(q = "web scraping", hl = "en")
search <- search %>% html_form_set(!!!vals)

#> Warning: Setting value of hidden field 'hl'.

# To submit and get result:
if (FALSE) {
session_submit(session, form)
}