Parse an html table into a data frame.

Parse an html table into a data frame.

html_table(x, header = NA, trim = TRUE, fill = FALSE, dec = ".")

Arguments

x
A node, node set or document.

header
Use first row as header? If NA, will use first row if it consists of <th> tags.

trim
Remove leading and trailing whitespace within each cell?

fill
If TRUE, automatically fill rows with fewer than the maximum number of columns with NAs.

dec
The character used as decimal mark.

Assumptions

html_table currently makes a few assumptions:

No cells span multiple rows

Headers are in the first row

Examples

tdist <- read_html("http://en.wikipedia.org/wiki/Student%27s_t-distribution")
tdist %>%
  html_node("table.infobox") %>%
  html_table(header = FALSE)

#>                                  X1
#> 1      Probability density function
#> 2  Cumulative distribution function
#> 3                        Parameters
#> 4                           Support
#> 5                               PDF
#> 6                               CDF
#> 7                              Mean
#> 8                            Median
#> 9                              Mode
#> 10                         Variance
#> 11                         Skewness
#> 12                     Ex. kurtosis
#> 13                          Entropy
#> 14                              MGF
#> 15                               CF
#>                                                                                                                                                                                                                                                                                                                                                                                                               X2
#> 1                                                                                                                                                                                                                                                                                                                                                                                   Probability density function
#> 2                                                                                                                                                                                                                                                                                                                                                                               Cumulative distribution function
#> 3                                                                                                                                                                                                                                                                                                                                                          ν>0{\\displaystyle \\nu >0} degrees of freedom (real)
#> 4                                                                                                                                                                                                                                                                                                                                                            x∈(−∞,∞){\\displaystyle x\\in (-\\infty ,\\infty )}
#> 5                                                                                                                                                         Γ(ν+12)νπΓ(ν2)(1+x2ν)−ν+12{\\displaystyle \\textstyle {\\frac {\\Gamma \\left({\\frac {\\nu +1}{2}}\\right)}{{\\sqrt {\\nu \\pi }}\\,\\Gamma \\left({\\frac {\\nu }{2}}\\right)}}\\left(1+{\\frac {x^{2}}{\\nu }}\\right)^{-{\\frac {\\nu +1}{2}}}\\!}
#> 6  12+xΓ(ν+12)×2F1(12,ν+12;32;−x2ν)πνΓ(ν2){\\displaystyle {\\begin{matrix}{\\frac {1}{2}}+x\\Gamma \\left({\\frac {\\nu +1}{2}}\\right)\\times \\\\[0.5em]{\\frac {\\,_{2}F_{1}\\left({\\frac {1}{2}},{\\frac {\\nu +1}{2}};{\\frac {3}{2}};-{\\frac {x^{2}}{\\nu }}\\right)}{{\\sqrt {\\pi \\nu }}\\,\\Gamma \\left({\\frac {\\nu }{2}}\\right)}}\\end{matrix}}}\n     where 2F1 is the hypergeometric function
#> 7                                                                                                                                                                                                                                                                                                                                                         0 for ν>1{\\displaystyle \\nu >1}, otherwise undefined
#> 8                                                                                                                                                                                                                                                                                                                                                                                                              0
#> 9                                                                                                                                                                                                                                                                                                                                                                                                              0
#> 10                                                                                                                                                                                                                                                   νν−2{\\displaystyle \\textstyle {\\frac {\\nu }{\\nu -2}}} for ν>2{\\displaystyle \\nu >2}, ∞ for 1<ν≤2{\\displaystyle 1<\\nu \\leq 2}, otherwise undefined
#> 11                                                                                                                                                                                                                                                                                                                                                        0 for ν>3{\\displaystyle \\nu >3}, otherwise undefined
#> 12                                                                                                                                                                                                                                                       6ν−4{\\displaystyle \\textstyle {\\frac {6}{\\nu -4}}} for ν>4{\\displaystyle \\nu >4}, ∞ for 2<ν≤4{\\displaystyle 2<\\nu \\leq 4}, otherwise undefined
#> 13                 ν+12[ψ(1+ν2)−ψ(ν2)]+ln⁡[νB(ν2,12)](nats){\\displaystyle {\\begin{matrix}{\\frac {\\nu +1}{2}}\\left[\\psi \\left({\\frac {1+\\nu }{2}}\\right)-\\psi \\left({\\frac {\\nu }{2}}\\right)\\right]\\\\[0.5em]+\\ln {\\left[{\\sqrt {\\nu }}B\\left({\\frac {\\nu }{2}},{\\frac {1}{2}}\\right)\\right]}\\,{\\scriptstyle {\\text{(nats)}}}\\end{matrix}}}\nψ: digamma function,\nB: beta function
#> 14                                                                                                                                                                                                                                                                                                                                                                                                     undefined
#> 15                                                                                  Kν/2(ν|t|)⋅(ν|t|)ν/2Γ(ν/2)2ν/2−1{\\displaystyle \\textstyle {\\frac {K_{\\nu /2}\\left({\\sqrt {\\nu }}|t|\\right)\\cdot \\left({\\sqrt {\\nu }}|t|\\right)^{\\nu /2}}{\\Gamma (\\nu /2)2^{\\nu /2-1}}}} for ν>0{\\displaystyle \\nu >0}\nKν(x){\\displaystyle K_{\\nu }(x)}: modified Bessel function of the second kind[1]

births <- read_html("https://www.ssa.gov/oact/babynames/numberUSbirths.html")
html_table(html_nodes(births, "table")[[2]])

#>     Year ofbirth      Male    Female     Total
#> 1           1880   118,400    97,605   216,005
#> 2           1881   108,282    98,855   207,137
#> 3           1882   122,031   115,695   237,726
#> 4           1883   112,477   120,059   232,536
#> 5           1884   122,738   137,586   260,324
#> 6           1885   115,945   141,948   257,893
#> 7           1886   119,041   153,735   272,776
#> 8           1887   109,313   155,422   264,735
#> 9           1888   129,906   189,445   319,351
#> 10          1889   119,032   189,219   308,251
#> 11          1890   119,701   201,661   321,362
#> 12          1891   109,265   196,566   305,831
#> 13          1892   131,453   224,913   356,366
#> 14          1893   121,041   225,232   346,273
#> 15          1894   124,893   235,971   360,864
#> 16          1895   126,643   247,106   373,749
#> 17          1896   129,071   251,993   381,064
#> 18          1897   121,942   248,275   370,217
#> 19          1898   132,104   274,144   406,248
#> 20          1899   115,193   247,490   362,683
#> 21          1900   162,133   317,752   479,885
#> 22          1901   115,595   254,229   369,824
#> 23          1902   132,750   280,334   413,084
#> 24          1903   129,326   278,196   407,522
#> 25          1904   138,508   292,436   430,944
#> 26          1905   143,238   309,865   453,103
#> 27          1906   144,071   313,440   457,511
#> 28          1907   158,587   337,434   496,021
#> 29          1908   166,368   354,531   520,899
#> 30          1909   176,867   368,096   544,963
#> 31          1910   208,521   419,526   628,047
#> 32          1911   241,395   441,818   683,213
#> 33          1912   451,457   586,712 1,038,169
#> 34          1913   536,247   654,913 1,191,160
#> 35          1914   683,325   796,619 1,479,944
#> 36          1915   880,940 1,023,876 1,904,816
#> 37          1916   923,257 1,085,721 2,008,978
#> 38          1917   959,325 1,123,698 2,083,023
#> 39          1918 1,048,674 1,202,365 2,251,039
#> 40          1919 1,015,337 1,174,648 2,189,985
#> 41          1920 1,100,817 1,244,039 2,344,856
#> 42          1921 1,137,940 1,279,711 2,417,651
#> 43          1922 1,125,261 1,247,556 2,372,817
#> 44          1923 1,132,325 1,252,460 2,384,785
#> 45          1924 1,169,066 1,295,706 2,464,772
#> 46          1925 1,151,445 1,263,075 2,414,520
#> 47          1926 1,145,506 1,230,152 2,375,658
#> 48          1927 1,161,855 1,236,397 2,398,252
#> 49          1928 1,141,171 1,195,463 2,336,634
#> 50          1929 1,107,538 1,157,521 2,265,059
#> 51          1930 1,129,422 1,166,423 2,295,845
#> 52          1931 1,069,568 1,103,625 2,173,193
#> 53          1932 1,074,289 1,106,222 2,180,511
#> 54          1933 1,020,008 1,045,936 2,065,944
#> 55          1934 1,061,738 1,082,225 2,143,963
#> 56          1935 1,069,387 1,086,706 2,156,093
#> 57          1936 1,064,170 1,077,491 2,141,661
#> 58          1937 1,093,483 1,101,762 2,195,245
#> 59          1938 1,136,302 1,141,359 2,277,661
#> 60          1939 1,133,162 1,134,043 2,267,205
#> 61          1940 1,186,022 1,181,249 2,367,271
#> 62          1941 1,254,654 1,245,857 2,500,511
#> 63          1942 1,408,070 1,390,394 2,798,464
#> 64          1943 1,454,311 1,435,274 2,889,585
#> 65          1944 1,388,968 1,366,455 2,755,423
#> 66          1945 1,371,285 1,346,064 2,717,349
#> 67          1946 1,650,220 1,612,845 3,263,065
#> 68          1947 1,857,382 1,817,823 3,675,205
#> 69          1948 1,782,564 1,742,576 3,525,140
#> 70          1949 1,801,835 1,755,492 3,557,327
#> 71          1950 1,819,067 1,758,618 3,577,685
#> 72          1951 1,911,828 1,847,277 3,759,105
#> 73          1952 1,974,199 1,902,276 3,876,475
#> 74          1953 2,001,256 1,929,099 3,930,355
#> 75          1954 2,068,413 1,990,839 4,059,252
#> 76          1955 2,089,594 2,004,694 4,094,288
#> 77          1956 2,144,575 2,059,267 4,203,842
#> 78          1957 2,187,459 2,097,547 4,285,006
#> 79          1958 2,153,039 2,065,020 4,218,059
#> 80          1959 2,166,256 2,078,511 4,244,767
#> 81          1960 2,165,696 2,079,841 4,245,537
#> 82          1961 2,155,956 2,076,400 4,232,356
#> 83          1962 2,102,051 2,026,983 4,129,034
#> 84          1963 2,065,187 1,987,940 4,053,127
#> 85          1964 2,027,330 1,957,248 3,984,578
#> 86          1965 1,895,210 1,827,379 3,722,589
#> 87          1966 1,817,912 1,755,580 3,573,492
#> 88          1967 1,779,774 1,716,699 3,496,473
#> 89          1968 1,776,069 1,709,549 3,485,618
#> 90          1969 1,829,979 1,762,722 3,592,701
#> 91          1970 1,905,622 1,831,889 3,737,511
#> 92          1971 1,818,381 1,752,365 3,570,746
#> 93          1972 1,674,675 1,612,464 3,287,139
#> 94          1973 1,614,211 1,554,035 3,168,246
#> 95          1974 1,630,740 1,566,138 3,196,878
#> 96          1975 1,623,029 1,560,738 3,183,767
#> 97          1976 1,633,225 1,571,864 3,205,089
#> 98          1977 1,709,859 1,644,926 3,354,785
#> 99          1978 1,709,099 1,643,789 3,352,888
#> 100         1979 1,791,825 1,723,133 3,514,958
#> 101         1980 1,855,024 1,780,361 3,635,385
#> 102         1981 1,862,416 1,788,234 3,650,650
#> 103         1982 1,887,126 1,813,985 3,701,111
#> 104         1983 1,863,331 1,789,354 3,652,685
#> 105         1984 1,876,225 1,802,890 3,679,115
#> 106         1985 1,923,822 1,845,903 3,769,725
#> 107         1986 1,920,958 1,845,046 3,766,004
#> 108         1987 1,949,471 1,873,831 3,823,302
#> 109         1988 2,001,329 1,922,580 3,923,909
#> 110         1989 2,095,455 1,991,890 4,087,345
#> 111         1990 2,151,160 2,053,962 4,205,122
#> 112         1991 2,119,192 2,033,116 4,152,308
#> 113         1992 2,098,620 2,004,297 4,102,917
#> 114         1993 2,064,960 1,971,175 4,036,135
#> 115         1994 2,037,946 1,949,024 3,986,970
#> 116         1995 2,010,995 1,921,218 3,932,213
#> 117         1996 2,003,372 1,916,882 3,920,254
#> 118         1997 1,997,337 1,908,808 3,906,145
#> 119         1998 2,027,136 1,938,008 3,965,144
#> 120         1999 2,038,346 1,946,204 3,984,550
#> 121         2000 2,087,391 1,994,881 4,082,272
#> 122         2001 2,067,422 1,980,040 4,047,462
#> 123         2002 2,065,531 1,973,957 4,039,488
#> 124         2003 2,100,052 2,005,406 4,105,458
#> 125         2004 2,112,190 2,016,585 4,128,775
#> 126         2005 2,126,034 2,028,174 4,154,208
#> 127         2006 2,190,837 2,089,049 4,279,886
#> 128         2007 2,213,423 2,114,797 4,328,220
#> 129         2008 2,178,453 2,080,929 4,259,382
#> 130         2009 2,118,953 2,022,977 4,141,930
#> 131         2010 2,052,122 1,958,284 4,010,406
#> 132         2011 2,028,589 1,935,047 3,963,636
#> 133         2012 2,025,890 1,936,239 3,962,129
#> 134         2013 2,016,731 1,923,856 3,940,587
#> 135         2014 2,044,378 1,951,652 3,996,030
#> 136         2015 2,038,331 1,945,317 3,983,648
#> 137         2016 2,017,790 1,928,438 3,946,228
#> 138         2017 1,963,290 1,874,899 3,838,189

# If the table is badly formed, and has different number of rows in
# each column use fill = TRUE. Here's it's due to incorrect colspan
# specification.
skiing <- read_html("http://data.fis-ski.com/dynamic/results.html?sector=CC&raceid=22395")
skiing %>%
  html_table(fill = TRUE)

#> list()