Functions renamed in rvest 1.0.0

rvest 1.0.0 renamed a number of functions to ensure that every function has a common prefix, matching tidyverse conventions that emerged since rvest was first created.

set_values() -> html_form_set()

submit_form() -> session_submit()

xml_tag() -> html_name()

xml_node() & html_node() -> html_element()

xml_nodes() & html_nodes() -> html_elements()

(html_node() and html_nodes() are only superseded because they're so widely used.)

Additionally all session related functions gained a common prefix:

html_session() -> session()

forward() -> session_forward()

back() -> session_back()

jump_to() -> session_jump_to()

follow_link() -> session_follow_link()

set_values(form, ...)

submit_form(session, form, submit = NULL, ...)

xml_tag(x)

xml_node(...)

xml_nodes(...)

html_nodes(...)

html_node(...)

back(x)

forward(x)

jump_to(x, url, ...)

follow_link(x, ...)

html_session(url, ...)