Work with xml.

Deprecated. Please use just xml2 directly

xml(x, ..., encoding = "")

xml_tag(x)

xml_node(x, css, xpath)

xml_nodes(x, css, xpath)

Arguments

x
A url, a local path, a string containing html, or a response from an httr request.

...
If x is a URL, additional arguments are passed on to httr::GET().

encoding
Specify encoding of document. See iconvlist() for complete list. If you have problems determining the correct encoding, try stringi::stri_enc_detect()

css
Nodes to select. Supply one of css or xpath depending on whether you want to use a CSS or XPath 1.0 selector.

xpath
Nodes to select. Supply one of css or xpath depending on whether you want to use a CSS or XPath 1.0 selector.